refactor: String을 NumberFormatting을 해주는 함수 String Extension으로 추가
